# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExampleSms API",
        "description": "This is a sample ExampleSms API server.  You can find out more about Swagger at",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "ExampleSms API v1.0.0 server"
        }
    ],
    "paths": {
...
        "/sms-report": {
            "get": {
                "tags": [
                    "sms"
                ],
                "summary": "SMS Report",
                "description": "SMS report endpoint.",
                "operationId": "reportSms",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2021-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2021-01-01"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "enum": [
                                "pending",
                                "delivered",
                                "failed"
                            ]
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "enum": [
                                "user_id",
                                "phone",
                                "send_time"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The SMS list has been fetched successfully."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{\n     *     'id': 1,\n     *     'phone': '905555555555',\n     *     'message': 'Hello World',\n     *     'origin': 'SMS',\n     *     'status': 'delivered',\n     *     'request': {\n     *     rquestData\n     *      },\n     *     'response': {\n     *     responseData\n     *     },\n     *     'send_time': '2021-01-01 00:00:00',\n     *     'request_time': '2021-01-01 00:00:00',\n     *     }"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": "{\n     *     'total': 1,\n     *     'per_page': 10,\n     *     'current_page': 1,\n     *     'last_page': 1,\n     *     'from': 1,\n     *     'to': 1,\n     *     }"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMS List not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sms-report-detail": {
            "get": {
                "tags": [
                    "sms"
                ],
                "summary": "SMS Report Detail",
                "description": "SMS report detail endpoint.",
                "operationId": "reportSmsDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "SMS ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMS found."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{\n     * 'id': 1,\n     * 'phone': '905555555555',\n     * 'message': 'Hello World',\n     * 'origin': 'SMS',\n     * 'status': 'delivered',\n     * 'request': {\n     * rquestData\n     * },\n     * 'response': {\n     * responseData\n     * },\n     * 'send_time': '2021-01-01 00:00:00',\n     * 'request_time': '2021-01-01 00:00:00',\n     * }"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMS not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}